requirements.txt
discord.py[voice]==2.3.2
wavelink==3.0.0
yt-dlp==2024.12.13
spotipy==2.24.0
python-dotenv==1.0.0
aiohttp==3.9.3
aiofiles==23.2.1